+++ libhoomd/computes/EvaluatorWalls.h
@@ -57,11 +57,7 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #define __EVALUATOR_WALLS_H__
 
 #ifndef NVCC
-#include <boost/shared_ptr.hpp>
-#include <boost/python.hpp>
-#include <boost/bind.hpp>
-
-#include "PotentialExternal.h"
+#include <string>
 #endif
 
 #include "BoxDim.h"
@@ -82,13 +78,13 @@ const unsigned int MAX_N_CWALLS=20;
 const unsigned int MAX_N_PWALLS=60;
 
 struct wall_type{
+    unsigned int     numSpheres; // these data types come first, since the structs are aligned already
+    unsigned int     numCylinders;
+    unsigned int     numPlanes;
     SphereWall       Spheres[MAX_N_SWALLS];
     CylinderWall     Cylinders[MAX_N_CWALLS];
     PlaneWall        Planes[MAX_N_PWALLS];
-    unsigned int     numSpheres;
-    unsigned int     numCylinders;
-    unsigned int     numPlanes;
-} __attribute__((aligned(8)));   // align on a 8 byte boundary so the GPU kernel can load it in 4 bytes at a time
+} __attribute__((aligned(4))); // align according to first member size
 
 //! Applys a wall force from all walls in the field parameter
 /*! \ingroup computes
@@ -164,9 +160,19 @@ class EvaluatorWalls
 
             if (evaluated)
                 {
+                // correctly result in a 0 force in this case
+                #ifdef NVCC
+                if (!isfinite(force_divr))
+                #else
+                if (!std::isfinite(force_divr))
+                #endif
+                    {
+                        force_divr = Scalar(0.0);
+                        pair_eng = Scalar(0.0);
+                    }
                 // add the force and potential energy to the particle i
                 F += dr*force_divr;
-                energy = pair_eng; // removing half since the other "particle" won't be represented * Scalar(0.5);
+                energy += pair_eng; // removing half since the other "particle" won't be represented * Scalar(0.5);
                 }
             }
 
@@ -188,9 +194,21 @@ class EvaluatorWalls
             if (evaluated)
                 {
                 // add the force and potential energy to the particle i
-                energy = pair_eng + force_divr * m_params.rextrap * r; // removing half since the other "particle" won't be represented * Scalar(0.5);
+                pair_eng = pair_eng + force_divr * m_params.rextrap * r; // removing half since the other "particle" won't be represented * Scalar(0.5);
                 force_divr *= m_params.rextrap / r;
-                F += dr * force_divr;
+                // correctly result in a 0 force in this case
+                #ifdef NVCC
+                if (!isfinite(force_divr))
+                #else
+                if (!std::isfinite(force_divr))
+                #endif
+                    {
+                        force_divr = Scalar(0.0);
+                        pair_eng = Scalar(0.0);
+                    }
+                // add the force and potential energy to the particle i
+                F += dr*force_divr;
+                energy += pair_eng; // removing half since the other "particle" won't be represented * Scalar(0.5);
                 }
             }
 
@@ -335,7 +353,7 @@ class EvaluatorWalls
         */
         static std::string getName()
             {
-            return std::string("walls_") + evaluator::getName();
+            return std::string("wall_") + evaluator::getName();
             }
         #endif
 
@@ -356,33 +374,5 @@ typename EvaluatorWalls<evaluator>::param_type make_wall_params(typename evaluat
     params.rextrap = rextrap;
     return params;
     }
+
 #endif //__EVALUATOR__WALLS_H__
-#ifndef NVCC
-
-//! Exports helper function for parameters based on standard evaluators
-template< class evaluator >
-void export_wall_params_helpers()
-    {
-    using namespace boost::python;
-    class_<typename EvaluatorWalls<evaluator>::param_type , boost::shared_ptr<typename EvaluatorWalls<evaluator>::param_type> >((EvaluatorWalls<evaluator>::getName()+"_params").c_str(), init<>())
-        .def_readwrite("params", &EvaluatorWalls<evaluator>::param_type::params)
-        .def_readwrite("rextrap", &EvaluatorWalls<evaluator>::param_type::rextrap)
-        .def_readwrite("rcutsq", &EvaluatorWalls<evaluator>::param_type::rcutsq)
-        ;
-    def(std::string("make_"+EvaluatorWalls<evaluator>::getName()+"_params").c_str(), &make_wall_params<evaluator>);
-    }
-
-//! Combines exports of evaluators and parameter helper functions
-template < class evaluator >
-void export_PotentialExternalWall(const std::string& name)
-    {
-    export_PotentialExternal< PotentialExternal<EvaluatorWalls<evaluator> > >(name);
-    export_wall_params_helpers<evaluator>();
-    }
-
-//! Helper function for converting python wall group structure to wall_type
-wall_type make_wall_field_params(boost::python::object walls, boost::shared_ptr<const ExecutionConfiguration> m_exec_conf);
-
-//! Exports walls helper function
-void export_wall_field_helpers();
-#endif