+++ libhoomd/updaters/TwoStepLangevinBase.cc
@@ -92,6 +92,13 @@ TwoStepLangevinBase::TwoStepLangevinBase(boost::shared_ptr<SystemDefinition> sys
     ArrayHandle<Scalar> h_gamma(m_gamma, access_location::host, access_mode::overwrite);
     for (unsigned int i = 0; i < m_gamma.size(); i++)
         h_gamma.data[i] = Scalar(1.0);
+        
+    // allocate memory for the per-type gamma_r storage and initialize them to 0.0 (no rotational noise by default)
+    GPUVector<Scalar> gamma_r(m_pdata->getNTypes(), m_exec_conf);
+    m_gamma_r.swap(gamma_r);
+    ArrayHandle<Scalar> h_gamma_r(m_gamma_r, access_location::host, access_mode::overwrite);
+    for (unsigned int i = 0; i < m_gamma_r.size(); i++)
+        h_gamma_r.data[i] = Scalar(1.0);
 
     // connect to the ParticleData to receive notifications when the maximum number of particles changes
     m_num_type_change_connection = m_pdata->connectNumTypesChange(boost::bind(&TwoStepLangevinBase::slotNumTypesChange, this));
@@ -107,16 +114,23 @@ void TwoStepLangevinBase::slotNumTypesChange()
     {
     // skip the reallocation if the number of types does not change
     // this keeps old parameters when restoring a snapshot
-    // it will result in invalid coeficients if the snapshot has a different type id -> name mapping
+    // it will result in invalid coefficients if the snapshot has a different type id -> name mapping
     if (m_pdata->getNTypes() == m_gamma.size())
         return;
 
     // re-allocate memory for the per-type gamma storage and initialize them to 1.0
     unsigned int old_ntypes = m_gamma.size();
     m_gamma.resize(m_pdata->getNTypes());
+    m_gamma_r.resize(m_pdata->getNTypes());
+    
     ArrayHandle<Scalar> h_gamma(m_gamma, access_location::host, access_mode::readwrite);
+    ArrayHandle<Scalar> h_gamma_r(m_gamma_r, access_location::host, access_mode::readwrite);
+    
     for (unsigned int i = old_ntypes; i < m_gamma.size(); i++)
+        {
         h_gamma.data[i] = Scalar(1.0);
+        h_gamma_r.data[i] = Scalar(1.0);
+        }
     }
 
 /*! \param typ Particle type to set gamma for
@@ -139,6 +153,28 @@ void TwoStepLangevinBase::setGamma(unsigned int typ, Scalar gamma)
     ArrayHandle<Scalar> h_gamma(m_gamma, access_location::host, access_mode::readwrite);
     h_gamma.data[typ] = gamma;
     }
+    
+    
+/*! \param typ Particle type to set gamma_r (2D rotational noise) for
+    \param gamma The gamma_r value to set
+*/    
+void TwoStepLangevinBase::setGamma_r(unsigned int typ, Scalar gamma_r)
+    {
+    // check for user errors
+    if (gamma_r < 0)
+        {
+        m_exec_conf->msg->error() << "gamma_r should be positive or 0! " << typ << endl;
+        throw runtime_error("Error setting params in TwoStepLangevinBase");
+        }
+    if (typ >= m_pdata->getNTypes())
+        {
+        m_exec_conf->msg->error() << "Trying to set gamma_r for a non existent type! " << typ << endl;
+        throw runtime_error("Error setting params in TwoStepLangevinBase");
+        }
+
+    ArrayHandle<Scalar> h_gamma_r(m_gamma_r, access_location::host, access_mode::readwrite);
+    h_gamma_r.data[typ] = gamma_r;
+    }
 
 void export_TwoStepLangevinBase()
     {
@@ -152,5 +188,6 @@ void export_TwoStepLangevinBase()
                                 >())
         .def("setT", &TwoStepLangevinBase::setT)
         .def("setGamma", &TwoStepLangevinBase::setGamma)
+        .def("setGamma_r", &TwoStepLangevinBase::setGamma_r)
         ;
     }