+++ libhoomd/updaters_gpu/ConstraintEllipsoidGPU.h
@@ -1,6 +1,6 @@
 /*
 Highly Optimized Object-oriented Many-particle Dynamics -- Blue Edition
-(HOOMD-blue) Open Source Software License Copyright 2009-2016 The Regents of
+(HOOMD-blue) Open Source Software License Copyright 2009-2015 The Regents of
 the University of Michigan All rights reserved.
 
 HOOMD-blue may contain modifications ("Contributions") provided, and to which
@@ -47,25 +47,43 @@ OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
+// Maintainer: joaander
 
-/*!
-\page page_compile_guide_linux_ubuntu Compiling HOOMD-blue in Ubuntu
+#include "ConstraintEllipsoid.h"
 
-This documentation was written and tested in **Ubuntu 12.04**. It may not work for other
-versions.
-
-\section software_req_install_ubuntu_dev Development environment
-
-1. Follow [NVIDIA's instructions](http://docs.nvidia.com/cuda/cuda-getting-started-guide-for-linux/index.html#package-manager-installation) to install the NVIDIA CUDA deb repository.
-
-2. Install g++, boost headers, python headers, zlib headers, git, openmpi, cmake, and CUDA.
-~~~~
-$ sudo apt-get install g++ libboost-all-dev zlib1g-dev python-dev git-core openmpi-bin libopenmpi-dev cmake cuda
-~~~~
-
-\section  software_req_install_ubuntu_build Compile hoomd
-
-Ubuntu provides an experimental pytyon3 packages alongside python27, but it does not provide boost::python compatible with python3. To build hoomd on an ubuntu system with python3 installed, you must specify `-DPYTHON_EXECUTABLE=/usr/bin/python2` in the initial cmake command to force python2.
-
-See \ref sec_build_linux_generic_compile for instructions on compiling hoomd.
+/*! \file ConstraintEllipsoidGPU.h
+    \brief Declares a class for computing ellipsoid constraint forces
 */
+
+#ifdef NVCC
+#error This header cannot be compiled by nvcc
+#endif
+
+#ifndef __CONSTRAINT_ELLIPSOID_GPU_H__
+#define __CONSTRAINT_ELLIPSOID_GPU_H__
+
+//! Applys a constraint force to keep a group of particles on a Ellipsoid
+/*! \ingroup computes
+*/
+class ConstraintEllipsoidGPU : public ConstraintEllipsoid
+    {
+    public:
+        //! Constructs the compute
+        ConstraintEllipsoidGPU(boost::shared_ptr<SystemDefinition> sysdef,
+                         boost::shared_ptr<ParticleGroup> group,
+                         Scalar3 P,
+                         Scalar rx,
+                         Scalar ry,
+                         Scalar rz);
+
+    protected:
+        unsigned int m_block_size;  //!< block size to execute on the GPU
+
+        //! Take one timestep forward
+        virtual void update(unsigned int timestep);
+    };
+
+//! Exports the ConstraintEllipsoidGPU class to python
+void export_ConstraintEllipsoidGPU();
+
+#endif