+++ libhoomd/updaters_gpu/TwoStepBDGPU.cc
@@ -79,8 +79,10 @@ TwoStepBDGPU::TwoStepBDGPU(boost::shared_ptr<SystemDefinition> sysdef,
                            boost::shared_ptr<Variant> T,
                            unsigned int seed,
                            bool use_lambda,
-                           Scalar lambda)
-    : TwoStepBD(sysdef, group, T, seed, use_lambda, lambda)
+                           Scalar lambda,
+                           bool noiseless_t,
+                           bool noiseless_r)
+    : TwoStepBD(sysdef, group, T, seed, use_lambda, lambda, noiseless_t, noiseless_r)
     {
     if (!m_exec_conf->isCUDAEnabled())
         {
@@ -116,6 +118,13 @@ void TwoStepBDGPU::integrateStepOne(unsigned int timestep)
     ArrayHandle<Scalar> d_diameter(m_pdata->getDiameters(), access_location::device, access_mode::read);
     ArrayHandle<unsigned int> d_tag(m_pdata->getTags(), access_location::device, access_mode::read);
 
+    // for rotational noise
+    ArrayHandle<Scalar> d_gamma_r(m_gamma_r, access_location::device, access_mode::read);
+    ArrayHandle<Scalar4> d_orientation(m_pdata->getOrientationArray(), access_location::device, access_mode::readwrite);
+    ArrayHandle<Scalar4> d_torque(m_pdata->getNetTorqueArray(), access_location::device, access_mode::readwrite);
+    ArrayHandle<Scalar3> d_inertia(m_pdata->getMomentsOfInertiaArray(), access_location::device, access_mode::read);
+    ArrayHandle<Scalar4> d_angmom(m_pdata->getAngularMomentumArray(), access_location::device, access_mode::readwrite);
+
     unsigned int num_blocks = group_size / m_block_size + 1;
 
     langevin_step_two_args args;
@@ -131,6 +140,8 @@ void TwoStepBDGPU::integrateStepOne(unsigned int timestep)
     args.block_size = m_block_size;
     args.num_blocks = num_blocks;
     args.tally = false;
+    
+    bool aniso = m_aniso;
 
     // perform the update on the GPU
     gpu_brownian_step_one(d_pos.data,
@@ -142,9 +153,17 @@ void TwoStepBDGPU::integrateStepOne(unsigned int timestep)
                           d_index_array.data,
                           group_size,
                           d_net_force.data,
+                          d_gamma_r.data,
+                          d_orientation.data,
+                          d_torque.data,
+                          d_inertia.data,
+                          d_angmom.data,
                           args,
+                          aniso,
                           m_deltaT,
-                          D);
+                          D, 
+                          m_noiseless_t, 
+                          m_noiseless_r);
 
     if(m_exec_conf->isCUDAErrorCheckingEnabled())
         CHECK_CUDA_ERROR();
@@ -170,7 +189,8 @@ void export_TwoStepBDGPU()
                                boost::shared_ptr<Variant>,
                                unsigned int,
                                bool,
-                               Scalar
-                               >())
+                               Scalar,
+                               bool,
+                               bool>())
         ;
     }